Split sim_gnss.py
